Fix named arguments with IFTI
