+++ compiler/src/dmd/frontend.h
@@ -1513,6 +1513,7 @@ public:
     ScopeDsymbol* argsym;
     size_t hash;
     Array<Expression* >* fargs;
+    Array<Identifier* >* fnames;
     Array<TemplateInstance* >* deferred;
     Module* memberOf;
     TemplateInstance* tinst;
@@ -4585,7 +4586,7 @@ public:
     TRUST trust;
     PURE purity;
     int8_t inuse;
-    Array<Expression* >* fargs;
+    ArgumentList inferenceArguments;
     static TypeFunction* create(Array<Parameter* >* parameters, Type* treturn, uint8_t varargs, LINK linkage, StorageClass stc = 0);
     const char* kind() const override;
     TypeFunction* syntaxCopy() override;
